Add is_iso8601_date boolean
Validate that a date being sent in as a string is a valid ISO 8601
date, meaning it matches YYYY-MM-DD

This was done in a somewhat generic way that we could support a
date_format option (and did for a time) allowing the user to specify
their own option, but honestly I think to do that right and provide
good default errors, we would need to take in a better date syntax
than Go's insane format.  Something like YYYY that users would
understand better than a random time like 2016-01-02.